Remove esm exports.
